backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
openai>=1.0.0
psycopg2-binary>=2.9.0
pinecone>=7.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
